set fit=True in sts.probplot
in scipy versions prior to 1.7.0 scipy.stats.probplot always plotted a line if the plot argument was not None. As we change the color of the plotted line afterwards this change in behaviour now causes an error in out code. This commit implements a quick workaround, by specifying fit=True as an argument to probplot so that the line is drawn in scipy 1.7.0 as well.